+++ src/layer/RLSecretLayer1.cpp
@@ -160,7 +160,7 @@ bool RLSecretLayer1::init() {
         m_crystalParticle->setVisible(false);
     }
 
-    if (rl::isUserOwner()) {
+    if (rl::isUserOwner() || rl::isUserDeveloper()) {
         // secret button for adding new codes
         auto secretSpr = CCSprite::createWithSpriteFrameName("RL_bigRuby.png"_spr);
         secretSpr->setScale(0.7f);